+++ packages/plugins/src/editors/communication/gse-editor.ts
@@ -15,7 +15,7 @@ import { newWizardEvent } from '@openscd/open-scd/src/foundation.js';
 import { newActionEvent } from '@openscd/core/foundation/deprecated/editor.js';
 import { sizableGooseIcon } from '@openscd/open-scd/src/icons/icons.js';
 import { editGseWizard } from '../../wizards/gse.js';
-import { getAllConnectedAPsOfSameIED } from './helpers.js';
+import { getAllConnectedAPsOfSameIED } from './foundation.js';
 
 @customElement('gse-editor')
 export class GseEditor extends LitElement {
